pyproject.toml
[tool.poetry]
name = "app_employee"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask-RESTful = "0.3.9"
Flask-SQLAlchemy = "2.5.1"
PyMySQL = "1.0.2"
marshmallow = "3.14.0"
flask-marshmallow = "0.14.0"
marshmallow-sqlalchemy = "0.26.1"
python-dotenv = "0.19.1"
gunicorn = "20.1.0"
gevent = "21.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
